# apidoc.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/carts/{id}": {
      "get": {
        "tags": [
          "cart-controller"
        ],
        "operationId": "getCartById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Cart"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "cart-controller"
        ],
        "operationId": "updateCart",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Cart"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Cart"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "cart-controller"
        ],
        "operationId": "deleteCart",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/carts": {
      "get": {
        "tags": [
          "cart-controller"
        ],
        "operationId": "getAllCarts",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageCart"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "cart-controller"
        ],
        "operationId": "createCart",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Cart"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Cart"
                }
              }
            }
          }
        }
      }
    },
...
    "/role/roleOf/{id}": {
      "get": {
        "tags": [
          "role-controller"
        ],
        "operationId": "getRole",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "uniqueItems": true,
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Role"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Role": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "users": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/User"
            }
          }
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "fullName": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "gender": {
            "type": "string",
            "enum": [
              "MALE",
              "FEMALE",
              "NOT_SPECIFIED"
            ]
          },
          "dob": {
            "type": "string",
            "format": "date"
          },
          "roles": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Role"
            }
          },
          "phoneNumber": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "profilePicture": {
            "type": "string"
          },
          "coverPicture": {
            "type": "string"
          },
          "about": {
            "type": "string"
          },
          "website": {
            "type": "string"
          },
          "facebook": {
            "type": "string"
          },
          "instagram": {
            "type": "string"
          },
          "twitter": {
            "type": "string"
          },
          "youtube": {
            "type": "string"
          },
          "github": {
            "type": "string"
          },
          "linkedin": {
            "type": "string"
          },
          "pinterest": {
            "type": "string"
          },
          "tiktok": {
            "type": "string"
          },
          "snapchat": {
            "type": "string"
          },
          "telegram": {
            "type": "string"
          },
          "whatsapp": {
            "type": "string"
          },
          "discord": {
            "type": "string"
          },
          "reddit": {
            "type": "string"
          }
        }
      },
      "Comment": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "commentText": {
            "type": "string"
          },
          "image": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Image"
            }
          },
          "video": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Video"
            }
          },
          "parent": {
            "$ref": "#/components/schemas/Comment"
          },
          "post": {
            "$ref": "#/components/schemas/Post"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          }
        }
      },
      "Image": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "path": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "imageBytes": {
            "type": "string",
            "format": "byte"
          },
          "description": {
            "type": "string"
          },
          "post": {
            "$ref": "#/components/schemas/Post"
          }
        }
      },
      "Post": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "creator": {
            "$ref": "#/components/schemas/User"
          },
          "postText": {
            "type": "string"
          },
          "postImage": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Image"
            }
          },
          "postVideo": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Video"
            }
          },
          "sharedPost": {
            "$ref": "#/components/schemas/Post"
          },
          "comments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Comment"
            }
          },
          "likers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/User"
            }
          },
          "shares": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/User"
            }
          },
          "views": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/User"
            }
          },
          "products": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Product"
            }
          },
          "public": {
            "type": "boolean"
          },
          "deleted": {
            "type": "boolean"
          },
          "archived": {
            "type": "boolean"
          }
        }
      },
      "Product": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "brand": {
            "type": "string"
          },
          "price": {
            "type": "number",
            "format": "double"
          },
          "discountPrice": {
            "type": "number",
            "format": "double"
          },
          "stockStatus": {
            "type": "string"
          },
          "quantity": {
            "type": "integer",
            "format": "int32"
          },
          "mainImage": {
            "$ref": "#/components/schemas/Image"
          },
          "galleryImages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Image"
            }
          },
          "dimensions": {
            "type": "string"
          },
          "weight": {
            "type": "string"
          },
          "seller": {
            "$ref": "#/components/schemas/User"
          },
          "ratings": {
            "type": "number",
            "format": "double"
          },
          "reviews": {
            "type": "integer",
            "format": "int32"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "origin": {
            "type": "string"
          },
          "features": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "addedDate": {
            "type": "string"
          }
        }
      },
...
      "Video": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "path": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "videoBytes": {
            "type": "string",
            "format": "byte"
          },
          "description": {
            "type": "string"
          },
          "post": {
            "$ref": "#/components/schemas/Post"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          }
        }
      },
...
      "Cart": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CartItem"
            }
          }
        }
      },
      "CartItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "product": {
            "$ref": "#/components/schemas/Product"
          },
          "quantity": {
            "type": "integer",
            "format": "int32"
          },
          "cart": {
            "$ref": "#/components/schemas/Cart"
          }
        }
      },
...
      "PageableObject": {
        "type": "object",
        "properties": {
          "offset": {
            "type": "integer",
            "format": "int64"
          },
          "sort": {
            "$ref": "#/components/schemas/SortObject"
          },
          "unpaged": {
            "type": "boolean"
          },
          "paged": {
            "type": "boolean"
          },
          "pageNumber": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "SortObject": {
        "type": "object",
        "properties": {
          "empty": {
            "type": "boolean"
          },
          "unsorted": {
            "type": "boolean"
          },
          "sorted": {
            "type": "boolean"
          }
        }
      },
...
      "PageCart": {
        "type": "object",
        "properties": {
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Cart"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "sort": {
            "$ref": "#/components/schemas/SortObject"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          },
          "numberOfElements": {
            "type": "integer",
            "format": "int32"
          },
          "pageable": {
            "$ref": "#/components/schemas/PageableObject"
          },
          "empty": {
            "type": "boolean"
          }
        }
      },
...
    }
  }
}