# src/main/java/com/jidnivai/sdcian/sdcian/entity/CartItem.java
package com.jidnivai.sdcian.sdcian.entity;

import com.fasterxml.jackson.annotation.JsonIgnore;

import jakarta.persistence.Entity;
import jakarta.persistence.GeneratedValue;
import jakarta.persistence.GenerationType;
import jakarta.persistence.Id;
import jakarta.persistence.ManyToOne;
import jakarta.persistence.OneToOne;
import lombok.Getter;
import lombok.Setter;
import lombok.ToString;

@Entity
@Getter
@Setter
@ToString
public class CartItem {
    @Id
    @GeneratedValue(strategy = GenerationType.AUTO)
    private Long id;
    
    @OneToOne
    private Product product;
    private int quantity;

    @ManyToOne
    @ToString.Exclude
    @JsonIgnore
    private Cart cart;
}